Separate UpsampleBase from upsample.h
Revert CMakeLists.txt to support CUDA 10.2 on Linux
