+++ onnxruntime/core/providers/cuda/tensor/upsample.h
@@ -5,7 +5,7 @@
 
 #include "core/providers/shared_library/provider_api.h"
 #include "core/providers/cuda/cuda_kernel.h"
-#include "core/providers/cpu/tensor/upsample.h"
+#include "core/providers/cpu/tensor/upsamplebase.h"
 
 namespace onnxruntime {
 namespace cuda {
