+++ onnxruntime/test/providers/cpu/tensor/upsample_op_test.cc
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/providers/cpu/tensor/upsample.h"
+#include "core/providers/cpu/tensor/upsamplebase.h"
 #include "gtest/gtest.h"
 #include "test/providers/provider_test_utils.h"
